requirements.txt
anthropic
python-telegram-bot
